Make user detail’s route public and return the data accordingly to the requester access level
