+++ politeiawww/www.go
@@ -1040,7 +1040,25 @@ func (p *politeiawww) handleUserDetails(w http.ResponseWriter, r *http.Request) 
 	var ud v1.UserDetails
 	ud.UserID = pathParams["userid"]
 
-	udr, err := p.backend.ProcessUserDetails(&ud)
+	userID, err := strconv.ParseUint(ud.UserID, 10, 64)
+	if err != nil {
+		RespondWithError(w, r, 0, "handleUserProposals: ParseUint",
+			v1.UserError{
+				ErrorCode: v1.ErrorStatusInvalidInput,
+			})
+		return
+	}
+
+	user, err := p.getSessionUser(r)
+	if err != nil {
+		log.Infof("handleUserDetails: could not get session user %v", err)
+	}
+
+	udr, err := p.backend.ProcessUserDetails(&ud,
+		user != nil && user.ID == userID,
+		user != nil && user.Admin,
+	)
+
 	if err != nil {
 		RespondWithError(w, r, 0,
 			"handleUserDetails: ProcessUserDetails %v", err)
@@ -1334,6 +1352,8 @@ func _main() error {
 		p.handleGetAllVoteStatus, permissionPublic, true)
 	p.addRoute(http.MethodGet, v1.RouteVoteStatus,
 		p.handleVoteStatus, permissionPublic, true)
+	p.addRoute(http.MethodGet, v1.RouteUserDetails,
+		p.handleUserDetails, permissionPublic, true)
 
 	// Routes that require being logged in.
 	p.addRoute(http.MethodPost, v1.RouteSecret, p.handleSecret,
@@ -1370,8 +1390,6 @@ func _main() error {
 		p.handleSetProposalStatus, permissionAdmin, true)
 	p.addRoute(http.MethodPost, v1.RouteStartVote,
 		p.handleStartVote, permissionAdmin, true)
-	p.addRoute(http.MethodGet, v1.RouteUserDetails,
-		p.handleUserDetails, permissionAdmin, true)
 	p.addRoute(http.MethodPost, v1.RouteEditUser,
 		p.handleEditUser, permissionAdmin, true)
